requirements.txt
alembic==1.8.1
anyio==3.6.1
autopep8==1.7.0
bcrypt==4.0.1
click==8.1.3
fastapi==0.85.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
passlib==1.7.4
pycodestyle==2.9.1
pydantic==1.10.2
pydantic-sqlalchemy==0.0.9
PyJWT==2.5.0
PyMySQL==1.0.2
python-decouple==3.6
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
toml==0.10.2
typing_extensions==4.4.0
uvicorn==0.18.3
